+++ trunk/linbox-old/src/library/objects/random/parametric/param_fuzzy_randiter.h
@@ -1,4 +1,4 @@
-/* File: src/library/objects/random/param_fuzzy_randiter.h
+/* File: src/library/objects/random/parametric/param_fuzzy_randiter.h
  * Author: William J Turner for the LinBox group
  */
 
@@ -97,7 +97,7 @@ namespace LinBox
       if (_size == 0)
 	return *(new element(rand()));
       else
-	return *(new element(static_cast<long>((double(rand())/RAND_MAX)*_size)));
+	return *(new element(static_cast<long>((double(rand())/RAND_MAX)*double(_size))));
     } // element& operator() (void)
 
   private:
